requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
kiteconnect>=4.2.0
pandas>=1.3.0
python-decouple>=3.6
sqlalchemy>=1.4.0
alembic>=1.7.0
mysqlclient>=2.0.3
pytest>=7.0.0
pytest-asyncio>=0.23.0 